Extracted EngineRequest from Request Passing $enableModifyingEngineBehaviorViaRequestParam
